pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
[project]
name = "dataforge-core"
version = "0.1.3"
authors = [
    {name="Vadim Orlov", email="vorlov@dataforgelabs.com"}
]
description = "Command line compiler for dataforge core projects"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml",
    "psycopg2-binary",
    "pyspark",
    "importlib_resources",
    "databricks-sql-connector",
    "sql-formatter"
    ]
[tool.setuptools.packages.find]
# All the following settings are optional:
where = ["cli"]  # ["."] by default
[tool.setuptools.package-data]
"*" = ["*.txt", "*.yaml", "*.sql"]
[project.scripts]
dataforge = "dataforge.main:main"
[project.urls]
Homepage = "https://github.com/dataforgelabs/dataforge-core"
Issues = "https://github.com/dataforgelabs/dataforge-core/issues"
